# cli: init plugin with boa and core