--- a/packages/cli/src/config.ts
+++ b/packages/cli/src/config.ts
@@ -4,7 +4,7 @@ export const dependencies: string[] = [
 ];
 
 export const daemonPackage = '@pipcook/daemon';
-export const boardPackage ='@pipcook/pipboard';
+export const boardPackage = '@pipcook/pipboard';
 
 export const pipcookLogName = 'pipcook-output';
 export const optionalNpmClients: string[] = [ 'npm', 'cnpm' ];