Speed up the plugin installation.

Benchmark:
Before: 1.16s user 0.26s system 0% cpu 4:08.73 total
```shell
$ time ./packages/cli/dist/bin/pipcook pipeline install ./example/pipelines/text-bayes-classification.json --tuna
ℹ Wrote to /path/.pipcook/plugins/package.json:
ℹ fetching package info @pipcook/plugins-csv-data-collect
ℹ {
ℹ   "name": "plugins",
ℹ   "version": "1.0.0",
ℹ   "description": "",
ℹ   "main": "index.js",
ℹ   "scripts": {
ℹ     "test": "echo \"Error: no test specified\" && exit 1"
ℹ   },
ℹ   "keywords": [],
ℹ   "author": "",
ℹ   "license": "ISC"
ℹ }
ℹ fetching package info @pipcook/plugins-csv-data-collect
ℹ fetching package info @pipcook/plugins-csv-data-collect
......
ℹ check plugin:
✔ @pipcook/plugins-csv-data-collect installed.
✔ @pipcook/plugins-csv-data-access installed.
✔ @pipcook/plugins-bayesian-model-define installed.
✔ @pipcook/plugins-bayesian-model-train installed.
✔ @pipcook/plugins-bayesian-model-evaluate installed.
✔ pipeline installed successfully.
./packages/cli/dist/bin/pipcook pipeline install  --tuna  1.16s user 0.26s system 0% cpu 4:08.73 total
```
After: 1.02s user 0.24s system 2% cpu 1:00.28 total
```shell
 time ./packages/cli/dist/bin/pipcook pipeline install ./example/pipelines/text-bayes-classification.json --tuna
ℹ fetching package info @pipcook/plugins-bayesian-model-evaluate
ℹ > @tensorflow/tfjs-node-gpu@1.7.0 install /path/.pipcook/plugins/node_modules/@tensorflow/tfjs-node-gpu
ℹ > node scripts/install.js gpu download
ℹ fetching package info @pipcook/plugins-bayesian-model-evaluate
ℹ GPU-darwin-1.7.0.tar.gz
⚠ * Downloading libtensorflow
⚠ fetching package info @pipcook/plugins-bayesian-model-evaluate
⚠ * Building TensorFlow Node.js bindings
ℹ fetching package info @pipcook/plugins-bayesian-model-evaluate
ℹ > nodejieba@2.4.1 install /path/.pipcook/plugins/node_modules/nodejieba
ℹ > node-pre-gyp install --fallback-to-build
ℹ fetching package info @pipcook/plugins-bayesian-model-evaluate
⚠ node-pre-gyp WARN Using request for node-pre-gyp https download 
⚠ node-pre-gyp WARN Tried to download(404): https://github.com/yanyiwu/nodejieba/releases/download/2.4.1/nodejieba-v2.4.1-node-v72-darwin-x64.tar.gz 
⚠ node-pre-gyp WARN Pre-built binaries not found for nodejieba@2.4.1 and node@12.17.0 (node-v72 ABI, unknown) (falling back to source compile with node-gyp) 
ℹ   CXX(target) Release/obj.target/nodejieba/lib/index.o
ℹ   CXX(target) Release/obj.target/nodejieba/lib/nodejieba.o
⚠ In file included from ../lib/nodejieba.cpp:3:
⚠ In file included from ../deps/cppjieba/Jieba.hpp:4:
⚠ In file included from ../deps/cppjieba/QuerySegment.hpp:8:
⚠ In file included from ../deps/cppjieba/DictTrie.hpp:13:
⚠ ../deps/limonp/StringUtil.hpp:87:70: warning: 'ptr_fun<unsigned int, bool>' is deprecated [-Wdeprecated-declarations]
⚠   s.erase(s.begin(), std::find_if(s.begin(), s.end(), std::not1(std::ptr_fun<unsigned, bool>(IsSpace))));
⚠                                                                      ^
⚠ /Library/Developer/CommandLineTools/usr/bin/../include/c++/v1/functional:1100:1: note: 'ptr_fun<unsigned int, bool>' has been explicitly marked deprecated here
⚠ _LIBCPP_DEPRECATED_IN_CXX11 inline _LIBCPP_INLINE_VISIBILITY
⚠ ^
⚠ /Library/Developer/CommandLineTools/usr/bin/../include/c++/v1/__config:1054:39: note: expanded from macro '_LIBCPP_DEPRECATED_IN_CXX11'
⚠ #  define _LIBCPP_DEPRECATED_IN_CXX11 _LIBCPP_DEPRECATED
⚠                                       ^
⚠ /Library/Developer/CommandLineTools/usr/bin/../include/c++/v1/__config:1043:48: note: expanded from macro '_LIBCPP_DEPRECATED'
......
ℹ check plugin:
✔ @pipcook/plugins-csv-data-collect installed.
✔ @pipcook/plugins-csv-data-access installed.
✔ @pipcook/plugins-bayesian-model-define installed.
✔ @pipcook/plugins-bayesian-model-train installed.
✔ @pipcook/plugins-bayesian-model-evaluate installed.
✔ pipeline installed successfully.
./packages/cli/dist/bin/pipcook pipeline install  --tuna  1.02s user 0.24s system 2% cpu 1:00.28 total
```