@yorkie 2020-08-19
It seems not too much improvement to speed up the plugin installation :(