@FeelyChau 2020-08-19
> It seems not too much improvement to speed up the plugin installation :(

1min from 4min, I think it's amazing improvement enough :)